+++ src/main/java/de/jeff_media/Drop2Inventory/Main.java
@@ -22,14 +22,16 @@ import org.bukkit.Material;
 
 public class Main extends JavaPlugin {
 
-	int currentConfigVersion = 10;
+	int currentConfigVersion = 12;
 
-	//BlockDropWrapper blockDropWrapper;
-	//DropHandler dropHandler;
+	BlockDropWrapper blockDropWrapper;
+	DropHandler dropHandler;
 	PluginUpdateChecker updateChecker;
 	Messages messages;
 	Utils utils;
+	MendingUtils mendingUtils;
 	IngotCondenser ingotCondenser;
+	ItemSpawnListener itemSpawnListener;
 
 	HashMap<String, PlayerSetting> perPlayerSettings;
 
@@ -60,9 +62,9 @@ public class Main extends JavaPlugin {
 		createConfig();
 
 		perPlayerSettings = new HashMap<String, PlayerSetting>();
-		//dropHandler = new DropHandler(this);
 		messages = new Messages(this);
 		ingotCondenser = new IngotCondenser(this);
+		itemSpawnListener = new ItemSpawnListener(this);
 		CommandDrop2Inv commandDrop2Inv = new CommandDrop2Inv(this);
 		
 		enabledByDefault = getConfig().getBoolean("enabled-by-default");
@@ -72,10 +74,20 @@ public class Main extends JavaPlugin {
 		autoCondense = getConfig().getBoolean("auto-condense");
 
 		this.getServer().getPluginManager().registerEvents(new Listener(this), this);
+		this.getServer().getPluginManager().registerEvents(itemSpawnListener,this);
+		if(mcVersion>=13) {
+			this.getServer().getPluginManager().registerEvents(new DropListener(this),this);
+			debug("MC Version is above 1.13, using BlockDropItemEvent");
+		} else {
+			blockDropWrapper = new BlockDropWrapper();
+			dropHandler = new DropHandler(this);
+			this.getServer().getPluginManager().registerEvents(new DropListenerLegacy(this),this);
+			debug("MC Version is 1.12 or below, using BlockBreakEvent");
+		}
 
 		utils = new Utils(this);
+		mendingUtils = new MendingUtils(this);
 
-		//blockDropWrapper = new BlockDropWrapper();
 		updateChecker = new PluginUpdateChecker(this,"https://api.jeff-media.de/drop2inventory/drop2inventory-latest-version.txt",
 				"https://www.spigotmc.org/resources/1-9-1-16-drop2inventory.62214/","https://github.com/JEFF-Media-GbR/Spigot-Drop2Inventory/blob/master/CHANGELOG.md","https://paypal.me/mfnalex");
 
@@ -184,6 +196,7 @@ public class Main extends JavaPlugin {
 		getConfig().addDefault("collect-block-exp", true);
 		getConfig().addDefault("collect-mob-exp", true);
 		getConfig().addDefault("auto-condense",false);
+		getConfig().addDefault("detect-legacy-drops",true);
 		disabledBlocks = new ArrayList<>();
 		disabledMobs = new ArrayList<>();
 		ArrayList<String> disabledBlocksStrings = (ArrayList<String>) getConfig().getStringList("disabled-blocks");
@@ -245,6 +258,8 @@ public class Main extends JavaPlugin {
 }
 	
 	boolean enabled(Player p) {
+
+		if(getConfig().getBoolean("always-enabled")) return true;
 		
 		// The following is for all the lazy server admins who use /reload instead of properly restarting their
 		// server ;) I am sometimes getting stacktraces although it is clearly stated that /reload is NOT