+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## 4.6.0
+- Added support for all properly coded custom drop plugins
+- Added support for plugins that create custom drops in ominous ways (toggleable)
+- Added possibility to use permissions per tool
+- Fixed "always-enabled" not working when the player disabled Drop2Inv before this setting was enabled
+
+## 4.5.2
+- Removed forgotten debug messages (sorry about that :( )
+
+## 4.5.1
+- Fixed Exception in MC 1.8
+- Fixed Exception when attempting to drop air
+- Changed event priority to support custom drops by other plugin
+
+## 4.5.0
+- Added limited compatibility for versions 1.8-1.12.2
+
 ## 4.4.1
 - Auto-Condense now also works for gold/iron nuggets, bone meal and slime balls. It also works when killing mobs now.
 - Fixed possible NullPointerException when Auto-Condense is enabled